Add files via upload
John the Ripper is a powerful tool in the cybersecurity capable of cracking a wide range of passwords and encrypted file formats. When used on password-protected files, it shows how vulnerable weak encryption can be. In a broader security assessment, it complements web vulnerability exploitation by revealing how poorly chosen or stored passwords can lead to serious breaches. 
where in this video password of encrypted zip file is cracked using John the Ripper tool in karl linux.